Reworked jdt-based incremental java compiler
Got rid of jdt compiler patches. Will use the same "magical" type
reference matching logic as implemented in (eclipse workspace)
IncrementalImageBuilder. This implies use of full classpath digest
to determine changes in external&reactor dependencies.

+++ takari-lifecycle-plugin/src/main/java/io/takari/maven/plugins/compile/jdt/ClassfileDigester.java
@@ -272,7 +272,7 @@ public class ClassfileDigester {
   }
 
   private void updateChar(char value) {
-    updateInt((int) value);
+    updateInt(value);
   }
 
   private void updateByte(byte value) {
